Schade Niederlande alles ist forbei
Fixed the camera, revamped the whole driving system. The target is now
reduced to a target which can be derived from the player position, as
it should be, not the other way around. The deriving caused the player
to be moved backwards when the robot has turned 180, while wanting to
move forward. The deriving wanted to correct the player position (and
move it backwards) before it started moving the player. Should now be
fixed.