Network discovery
Up and running, enter the command 'Discover' to see all players in your
subnet (excluding yourself :) ), enter the command 'Join' appended with
an integer denouncing the player with which you want to connect.

The code is mostly c-style, as reading the buffers was implemented in
net.h with char* so we want along with this.